requirements.txt
annotated-types==0.7.0
anyio==4.4.0
apt-clone==0.2.1
apturl==0.5.2
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
eyeD3==0.8.10
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.6.0
greenlet==1.1.2
grpcio==1.30.2
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
more-itertools==8.10.0
msgpack==1.0.3
nemo-emblems==5.8.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
onboard==1.4.1
orjson==3.10.6
packaging==21.3
PAM==0.4.2
pexpect==4.8.0
pika==1.3.2
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pydantic==2.8.2
pydantic_core==2.20.1
pyelftools==0.27
Pygments==2.18.0
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-magic==0.4.24
python-multipart==0.0.9
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
rich==13.7.1
SecretStorage==3.3.1
setproctitle==1.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
starlette==0.37.2
systemd-python==234
tinycss2==1.1.1
tldextract==3.1.2
typer==0.12.3
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.10.0
Unidecode==1.3.3
urllib3==1.26.5
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.13.0+ds
wadllib==1.3.6
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zipp==1.0.0
